+++ CHANGELOG.md
@@ -5,7 +5,13 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/)
 principles.
 
-## 3.10.0
+## [Unreleased]
+
+### Added
+
+- new field to `VOUCHER`: `redemption_count`
+
+## [3.10.0]
 
 ### Added